# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Career Advisor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List all skills
      responses:
        "200":
          description: List of skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
    post:
      operationId: createSkill
      tags: [skills]
      summary: Add a new skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSkillBody"
      responses:
        "201":
          description: Created skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /skills/{id}:
    put:
      operationId: updateSkill
      tags: [skills]
      summary: Update a skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSkillBody"
      responses:
        "200":
          description: Updated skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
    delete:
      operationId: deleteSkill
      tags: [skills]
      summary: Delete a skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Add a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Skill:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [technical, soft, language, tool, framework, other]
        proficiency:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - proficiency
        - createdAt
    CreateSkillBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [technical, soft, language, tool, framework, other]
        proficiency:
          type: string
          enum: [beginner, intermediate, advanced, expert]
      required:
        - name
        - category
        - proficiency
    UpdateSkillBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [technical, soft, language, tool, framework, other]
        proficiency:
          type: string
          enum: [beginner, intermediate, advanced, expert]
    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        techStack:
          type: string
        githubUrl:
          type: string
        liveUrl:
          type: string
        status:
          type: string
          enum: [in_progress, completed, paused]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - status
        - createdAt
    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        techStack:
          type: string
        githubUrl:
          type: string
        liveUrl:
          type: string
        status:
          type: string
          enum: [in_progress, completed, paused]
      required:
        - title
        - description
        - status
    UpdateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        techStack:
          type: string
        githubUrl:
          type: string
        liveUrl:
          type: string
        status:
          type: string
          enum: [in_progress, completed, paused]
...