# lib/api-zod/src/generated/types/createSkillBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Career Advisor API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSkillBodyCategory } from "./createSkillBodyCategory";
import type { CreateSkillBodyProficiency } from "./createSkillBodyProficiency";

export interface CreateSkillBody {
  name: string;
  category: CreateSkillBodyCategory;
  proficiency: CreateSkillBodyProficiency;
}