# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Career Advisor API
 * OpenAPI spec version: 0.1.0
 */
import type { SkillCategory } from "./skillCategory";
import type { SkillProficiency } from "./skillProficiency";

export interface Skill {
  id: number;
  name: string;
  category: SkillCategory;
  proficiency: SkillProficiency;
  createdAt: Date;
}