# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Career Advisor API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectBodyStatus } from "./updateProjectBodyStatus";

export interface UpdateProjectBody {
  title?: string;
  description?: string;
  techStack?: string;
  githubUrl?: string;
  liveUrl?: string;
  status?: UpdateProjectBodyStatus;
}